refactor: move the static methods for retrieving case files from FileSideBar to FilesProvider
