+++ src/main/java/school/hei/patrimoine/visualisation/swing/ihm/google/component/files/FileSideBar.java
@@ -1,15 +1,13 @@
 package school.hei.patrimoine.visualisation.swing.ihm.google.component.files;
 
-import static java.util.Objects.requireNonNull;
 import static school.hei.patrimoine.patrilang.PatriLangTranspiler.*;
 import static school.hei.patrimoine.visualisation.swing.ihm.google.config.EnvironmentConfig.isOfflineMode;
 import static school.hei.patrimoine.visualisation.swing.ihm.google.modele.FileCategory.*;
 import static school.hei.patrimoine.visualisation.swing.ihm.google.modele.GoogleLinkListDownloader.*;
+import static school.hei.patrimoine.visualisation.swing.ihm.google.providers.FilesProvider.*;
 
 import java.awt.*;
 import java.io.File;
-import java.util.Arrays;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import javax.swing.*;
@@ -27,16 +25,16 @@ public class FileSideBar extends JPanel {
     super(new BorderLayout());
 
     this.state = state;
-    this.plannedList = new JList<>(new FileListModel(getPatriLangPlannedFiles()));
     this.doneList = new JList<>(new FileListModel(getPatriLangDoneFiles()));
+    this.plannedList = new JList<>(new FileListModel(getPatriLangPlannedFiles()));
     this.justificativeList = new JList<>(new FileListModel(getPatriLangJustificativeFiles()));
 
-    plannedList.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
     doneList.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
+    plannedList.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
     justificativeList.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
 
-    plannedList.setCellRenderer(new FileListCellRenderer());
     doneList.setCellRenderer(new FileListCellRenderer());
+    plannedList.setCellRenderer(new FileListCellRenderer());
     justificativeList.setCellRenderer(new FileListCellRenderer());
 
     var plannedDebouncer =
@@ -147,20 +145,21 @@ public class FileSideBar extends JPanel {
 
     var plannedLabel = new JLabel("Journaux planifiés");
     plannedLabel.setFont(plannedLabel.getFont().deriveFont(Font.BOLD));
+
     panel.add(plannedLabel);
     panel.add(new JScrollPane(plannedList));
-
     panel.add(Box.createVerticalStrut(20));
 
     var doneLabel = new JLabel("Journaux réalisés");
     doneLabel.setFont(doneLabel.getFont().deriveFont(Font.BOLD));
+
     panel.add(doneLabel);
     panel.add(new JScrollPane(doneList));
-
     panel.add(Box.createVerticalStrut(20));
 
     var justificativeLabel = new JLabel("Pièces justificatives");
     justificativeLabel.setFont(justificativeLabel.getFont().deriveFont(Font.BOLD));
+
     panel.add(justificativeLabel);
     panel.add(new JScrollPane(justificativeList));
 
@@ -176,9 +175,9 @@ public class FileSideBar extends JPanel {
 
     var namedIds =
         switch (category) {
-          case JUSTIFICATIVE -> ids.justificative();
-          case PLANNED -> ids.planned();
           case DONE -> ids.done();
+          case PLANNED -> ids.planned();
+          case JUSTIFICATIVE -> ids.justificative();
         };
 
     return namedIds.stream()
@@ -187,56 +186,12 @@ public class FileSideBar extends JPanel {
               var filename =
                   currentFile
                       .getName()
-                      .replace(TOUT_CAS_FILE_EXTENSION, "")
+                      .replace(PJ_FILE_EXTENSION, "")
                       .replace(CAS_FILE_EXTENSION, "")
-                      .replace(PJ_FILE_EXTENSION, "");
+                      .replace(TOUT_CAS_FILE_EXTENSION, "");
               return driveNamedId.name().equals(filename);
             })
         .findFirst()
         .map(NamedID::id);
-  }
-
-  public static List<File> getPatriLangPlannedFiles() {
-    return Arrays.stream(requireNonNull(new File(getPlannedDirectoryPath()).listFiles()))
-        .filter(
-            file ->
-                file.getName().endsWith(TOUT_CAS_FILE_EXTENSION)
-                    || file.getName().endsWith(CAS_FILE_EXTENSION))
-        .toList();
-  }
-
-  public static List<File> getPatriLangDoneFiles() {
-    return Arrays.stream(requireNonNull(new File(getDoneDirectoryPath()).listFiles()))
-        .filter(
-            file ->
-                file.getName().endsWith(TOUT_CAS_FILE_EXTENSION)
-                    || file.getName().endsWith(CAS_FILE_EXTENSION))
-        .toList();
-  }
-
-  public static List<File> getPatriLangJustificativeFiles() {
-    return Arrays.stream(requireNonNull(new File(getJustificativeDirectoryPath()).listFiles()))
-        .filter(file -> file.getName().endsWith(PJ_FILE_EXTENSION))
-        .toList();
-  }
-
-  public static File getPlannedCasSetFile() {
-    return Arrays.stream(requireNonNull(new File(getPlannedDirectoryPath()).listFiles()))
-        .filter(file -> file.getName().endsWith(TOUT_CAS_FILE_EXTENSION))
-        .findFirst()
-        .orElseThrow();
-  }
-
-  public static File getDoneCasSetFile() {
-    return Arrays.stream(requireNonNull(new File(getDoneDirectoryPath()).listFiles()))
-        .filter(file -> file.getName().endsWith(TOUT_CAS_FILE_EXTENSION))
-        .findFirst()
-        .orElseThrow();
-  }
-
-  public static List<File> getDonePatrilangFilesWithoutCasSet() {
-    return getPatriLangDoneFiles().stream()
-        .filter(file -> !file.getName().endsWith(TOUT_CAS_FILE_EXTENSION))
-        .toList();
   }
 }
