+++ src/main/java/school/hei/patrimoine/visualisation/swing/ihm/google/component/recoupement/PossessionRecoupeeRealisationsDialog.java
@@ -4,6 +4,7 @@ import static java.util.stream.Collectors.joining;
 import static school.hei.patrimoine.modele.recouppement.model.RecoupementStatus.IMPREVU;
 import static school.hei.patrimoine.visualisation.swing.ihm.google.component.app.ViewFactory.make;
 import static school.hei.patrimoine.visualisation.swing.ihm.google.modele.MessageDialog.*;
+import static school.hei.patrimoine.visualisation.swing.ihm.google.providers.FilesProvider.getDoneCasSetFile;
 
 import java.awt.*;
 import java.io.File;
@@ -24,7 +25,6 @@ import school.hei.patrimoine.visualisation.swing.ihm.google.component.Dialog;
 import school.hei.patrimoine.visualisation.swing.ihm.google.component.app.AppContext;
 import school.hei.patrimoine.visualisation.swing.ihm.google.component.app.MultiViews;
 import school.hei.patrimoine.visualisation.swing.ihm.google.component.button.Button;
-import school.hei.patrimoine.visualisation.swing.ihm.google.component.files.FileSideBar;
 import school.hei.patrimoine.visualisation.swing.ihm.google.generator.PossessionGeneratorFactory;
 import school.hei.patrimoine.visualisation.swing.ihm.google.modele.AsyncTask;
 import school.hei.patrimoine.visualisation.swing.ihm.google.modele.State;
@@ -194,7 +194,7 @@ public class PossessionRecoupeeRealisationsDialog extends Dialog {
                 })
             .collect(joining("\n"));
 
-    File casSet = FileSideBar.getDoneCasSetFile();
+    File casSet = getDoneCasSetFile();
     File selectedFile = state.get("selectedFile");
 
     AsyncTask.<Void>builder()
